+++ fast/stages/03-project-factory/prod/variables.tf
@@ -41,6 +41,17 @@ variable "defaults_file" {
   default     = "data/defaults.yaml"
 }
 
+variable "prefix" {
+  # tfdoc:variable:source 00-bootstrap
+  description = "Prefix used for resources that need unique names. Use 9 characters or less."
+  type        = string
+
+  validation {
+    condition     = try(length(var.prefix), 0) < 10
+    error_message = "Use a maximum of 9 characters for prefix."
+  }
+}
+
 variable "shared_vpc_self_link" {
   # tfdoc:variable:source 02-networking
   description = "Self link for the shared VPC."