+++ tools/check_documentation.py
@@ -13,7 +13,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
 '''Recursively check freshness of tfdoc's generated tables in README files.
 
 This tool recursively checks that the embedded variables and outputs tables in
@@ -29,9 +28,7 @@ import pathlib
 import click
 import tfdoc
 
-
 BASEDIR = pathlib.Path(__file__).resolve().parents[1]
-
 
 State = enum.Enum('State', 'OK FAIL SKIP')
 
@@ -59,11 +56,9 @@ def _check_dir(dir_name, exclude_files=None, files=False, show_extra=False):
           state = State.OK
         else:
           state = State.FAIL
-          diff = '\n'.join(
-              [f'----- {mod_name} diff -----\n'] +
-              list(difflib.ndiff(
-                  result['doc'].split('\n'), new_doc.split('\n')
-              )))
+          header = f'----- {mod_name} diff -----\n'
+          ndiff = difflib.ndiff(result['doc'].split('\n'), new_doc.split('\n'))
+          diff = '\n'.join([header] + list(ndiff))
     yield mod_name, state, diff